Substituted screenshot of "dashboard" version of the demo
The Boston Car Data Explorer demo now has a small link next to the year
selector called “1920x1080”. If you click this, it launches the demo in
a new 1920x1080 browser window, and uses a “dashboard”-like version of
the demo which all fits on one page. This makes for a better screenshot.